Initial working program
A python program to construct word frequency lists from Wikipedia
pages. Input is a .xml file as downloaded from
https://en.wikipedia.org/wiki/Special:Export, the equivalent for
a different language or the complete xml structure for an entire
wiki, as downloaded from
https://dumps.wikimedia.org/backup-index.html

An additional optional parameter is the number of words in the
frequency list, with a default of 1000 words. If you just want the
top 10 words, for instance, it would look like this:
	python extract_wordlist.py filename.xml 10

Current issues are that it doesn't process serially, so it chokes
on large files. In addition, a couple markup/html keywords seem to
sneak through the filters.